yt-transcriptor/requirements.txt
youtube-transcript-api==0.6.1
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
beautifulsoup4==4.12.2
requests==2.31.0
newspaper3k==0.2.8
trafilatura==1.6.1
